# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Course Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterBody {
  name: string;
  email: string;
  password: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  role: string;
  approved: boolean;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface AdminUser {
  id: number;
  name: string;
  email: string;
  role: string;
  approved: boolean;
  createdAt: string;
  enrollmentCount: number;
}

export interface Course {
  id: number;
  title: string;
  description: string;
  coverImage: string;
  status: string;
  createdAt: string;
}

export interface CourseDetail {
  id: number;
  title: string;
  description: string;
  coverImage: string;
  status: string;
  createdAt: string;
  /** @nullable */
  contactEmail?: string | null;
  /** @nullable */
  contactPhone?: string | null;
  /** @nullable */
  whatsappNumber?: string | null;
  /** @nullable */
  instructorName?: string | null;
  enrolled: boolean;
}

export interface CreateCourseBody {
  title: string;
  description: string;
  coverImage: string;
  status: string;
  contactEmail?: string;
  contactPhone?: string;
  whatsappNumber?: string;
  instructorName?: string;
}

export interface UpdateCourseBody {
  title?: string;
  description?: string;
  coverImage?: string;
  status?: string;
  contactEmail?: string;
  contactPhone?: string;
  whatsappNumber?: string;
  instructorName?: string;
}

export interface EnrolledCourse {
  id: number;
  title: string;
  description: string;
  coverImage: string;
  status: string;
  progress: number;
  totalVideos: number;
  watchedVideos: number;
}

export interface Video {
  id: number;
  title: string;
  url: string;
  sortOrder: number;
  sectionId: number;
  watched?: boolean;
}

export interface VideoSection {
  id: number;
  title: string;
  sortOrder: number;
  courseId: number;
  videos: Video[];
}

export interface CourseFile {
  id: number;
  name: string;
  url: string;
  categoryId: number;
}

export interface FileCategory {
  id: number;
  name: string;
  courseId: number;
  files: CourseFile[];
}

export interface InteractiveSession {
  id: number;
  title: string;
  link: string;
  /** @nullable */
  scheduledAt?: string | null;
  courseId: number;
}

export interface CourseContact {
  /** @nullable */
  instructorName?: string | null;
  /** @nullable */
  contactEmail?: string | null;
  /** @nullable */
  contactPhone?: string | null;
  /** @nullable */
  whatsappNumber?: string | null;
}

export interface CourseProgress {
  courseId: number;
  totalVideos: number;
  watchedVideos: number;
  progress: number;
  watchedVideoIds: number[];
}

export interface Device {
  id: number;
  ip: string;
  userAgent: string;
  lastActive: string;
  current: boolean;
}

export interface EnrollUserBody {
  userId: number;
}

export interface EnrollmentInfo {
  userId: number;
  userName: string;
  userEmail: string;
  enrolledAt: string;
  progress: number;
}

export interface CreateSectionBody {
  title: string;
  sortOrder?: number;
}

export interface CreateVideoBody {
  title: string;
  url: string;
  sortOrder?: number;
}

export interface CreateFileCategoryBody {
  name: string;
}

export interface CreateFileBody {
  name: string;
  url: string;
}

export interface CreateSessionBody {
  title: string;
  link: string;
  scheduledAt?: string;
}

export interface DashboardSummary {
  enrolledCourses: number;
  totalProgress: number;
  totalVideosWatched: number;
  recentCourses: EnrolledCourse[];
}

export interface AdminDashboard {
  totalUsers: number;
  approvedUsers: number;
  pendingUsers: number;
  totalCourses: number;
  totalEnrollments: number;
}