# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Course Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Admin login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /admin/users/{id}/approve:
    patch:
      operationId: adminApproveUser
      tags: [admin]
      summary: Approve a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /admin/users/{id}/deny:
    patch:
      operationId: adminDenyUser
      tags: [admin]
      summary: Deny a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /admin/courses:
    get:
      operationId: adminListCourses
      tags: [admin]
      summary: List all courses (admin)
      responses:
        "200":
          description: Course list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      operationId: adminCreateCourse
      tags: [admin]
      summary: Create a course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourseBody"
      responses:
        "201":
          description: Course created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /admin/courses/{id}:
    patch:
      operationId: adminUpdateCourse
      tags: [admin]
      summary: Update a course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCourseBody"
      responses:
        "200":
          description: Course updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    delete:
      operationId: adminDeleteCourse
      tags: [admin]
      summary: Delete a course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/courses/{courseId}/enroll:
    post:
      operationId: adminEnrollUser
      tags: [admin]
      summary: Enroll a user in a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrollUserBody"
      responses:
        "201":
          description: User enrolled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/courses/{courseId}/enrollments:
    get:
      operationId: adminListEnrollments
      tags: [admin]
      summary: List enrolled users for a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Enrollment list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EnrollmentInfo"
  /admin/courses/{courseId}/sections:
    post:
      operationId: adminCreateSection
      tags: [admin]
      summary: Create a video section
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSectionBody"
      responses:
        "201":
          description: Section created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoSection"
  /admin/sections/{sectionId}/videos:
    post:
      operationId: adminCreateVideo
      tags: [admin]
      summary: Add a video to a section
      parameters:
        - name: sectionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVideoBody"
      responses:
        "201":
          description: Video added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
  /admin/courses/{courseId}/file-categories:
    post:
      operationId: adminCreateFileCategory
      tags: [admin]
      summary: Create a file category
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFileCategoryBody"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileCategory"
  /admin/file-categories/{categoryId}/files:
    post:
      operationId: adminCreateFile
      tags: [admin]
      summary: Add a file to a category
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFileBody"
      responses:
        "201":
          description: File added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseFile"
  /admin/courses/{courseId}/sessions:
    post:
      operationId: adminCreateSession
      tags: [admin]
      summary: Create an interactive session
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionBody"
      responses:
        "201":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InteractiveSession"
  /admin/videos/{id}:
    delete:
      operationId: adminDeleteVideo
      tags: [admin]
      summary: Delete a video
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Video deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/sections/{id}:
    delete:
      operationId: adminDeleteSection
      tags: [admin]
      summary: Delete a video section
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Section deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/files/{id}:
    delete:
      operationId: adminDeleteFile
      tags: [admin]
      summary: Delete a file
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: File deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/file-categories/{id}:
    delete:
      operationId: adminDeleteFileCategory
      tags: [admin]
      summary: Delete a file category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/sessions/{id}:
    delete:
      operationId: adminDeleteSession
      tags: [admin]
      summary: Delete a session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Session deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        approved:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - approved
        - createdAt
    AdminUser:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        approved:
          type: boolean
        createdAt:
          type: string
        enrollmentCount:
          type: integer
      required:
        - id
        - name
        - email
        - role
        - approved
        - createdAt
        - enrollmentCount
...
    Course:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        coverImage:
          type: string
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - title
        - description
        - coverImage
        - status
        - createdAt
...
    CreateCourseBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        coverImage:
          type: string
        status:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        whatsappNumber:
          type: string
        instructorName:
          type: string
        teacherId:
          type: integer
      required:
        - title
        - description
        - coverImage
        - status
    UpdateCourseBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        coverImage:
          type: string
        status:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        whatsappNumber:
          type: string
        instructorName:
          type: string
        teacherId:
          type: integer
...
    VideoSection:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        sortOrder:
          type: integer
        courseId:
          type: integer
        videos:
          type: array
          items:
            $ref: "#/components/schemas/Video"
      required:
        - id
        - title
        - sortOrder
        - courseId
        - videos
    Video:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        url:
          type: string
        sortOrder:
          type: integer
        sectionId:
          type: integer
        watched:
          type: boolean
      required:
        - id
        - title
        - url
        - sortOrder
        - sectionId
    FileCategory:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        courseId:
          type: integer
        files:
          type: array
          items:
            $ref: "#/components/schemas/CourseFile"
      required:
        - id
        - name
        - courseId
        - files
    CourseFile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        url:
          type: string
        categoryId:
          type: integer
      required:
        - id
        - name
        - url
        - categoryId
    InteractiveSession:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        link:
          type: string
        scheduledAt:
          type: ["string", "null"]
        courseId:
          type: integer
      required:
        - id
        - title
        - link
        - courseId
...
    EnrollUserBody:
      type: object
      properties:
        userId:
          type: integer
      required:
        - userId
    EnrollmentInfo:
      type: object
      properties:
        userId:
          type: integer
        userName:
          type: string
        userEmail:
          type: string
        enrolledAt:
          type: string
        progress:
          type: number
      required:
        - userId
        - userName
        - userEmail
        - enrolledAt
        - progress
    CreateSectionBody:
      type: object
      properties:
        title:
          type: string
        sortOrder:
          type: integer
      required:
        - title
    CreateVideoBody:
      type: object
      properties:
        title:
          type: string
        url:
          type: string
        sortOrder:
          type: integer
      required:
        - title
        - url
    CreateFileCategoryBody:
      type: object
      properties:
        name:
          type: string
      required:
        - name
    CreateFileBody:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
      required:
        - name
        - url
    CreateSessionBody:
      type: object
      properties:
        title:
          type: string
        link:
          type: string
        scheduledAt:
          type: string
      required:
        - title
        - link
...