# lib/api-zod/src/generated/types/course.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Course Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Course {
  id: number;
  title: string;
  description: string;
  coverImage: string;
  status: string;
  createdAt: string;
}