# lib/api-zod/src/generated/types/fileCategory.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Course Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CourseFile } from "./courseFile";

export interface FileCategory {
  id: number;
  name: string;
  courseId: number;
  files: CourseFile[];
}