# lib/api-zod/src/generated/types/interactiveSession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Course Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface InteractiveSession {
  id: number;
  title: string;
  link: string;
  /** @nullable */
  scheduledAt?: string | null;
  courseId: number;
}